requirements.txt
opencv-python
ultralytics
Roboflow
python-dotenv